apollo and beyound for better working querys and a server
