+++ backend/package.json
@@ -19,6 +19,7 @@
   "dependencies": {
     "@graphql-tools/schema": "^7.1.3",
     "@prisma/client": "^2.17.0",
+    "apollo-server": "^2.21.0",
     "apollo-server-express": "^2.21.0",
     "bcryptjs": "^2.4.3",
     "concurrently": "^6.0.0",
