+++ backend/schema.js
@@ -1,0 +1,18 @@
+import { gql } from "apollo-server";
+
+const typeDefs = gql`
+  type User {
+    email: String!
+    firstname: String
+  }
+  type Tweet {
+    id: ID!
+    description: String
+  }
+  type Query {
+    allUsers: [User!]!
+    allTweets: [Tweet!]!
+  }
+`;
+
+export default typeDefs;
