+++ frontend/src/App.js
@@ -5,7 +5,7 @@ import Routes from "./config/routes";
 
 import "./App.css";
 const client = new ApolloClient({
-  uri: "http://localhost:4000/graphql",
+  uri: "http://localhost:4000",
   cache: new InMemoryCache(),
 });
